--- src/core/index.ts
@@ -1,6 +1,0 @@
-// Audience Feed Connector
-export * from "./interfaces/mediarithmics/api/audience-feed-connector/AudienceFeedInterface"
-export * from "./interfaces/mediarithmics/api/audience-feed-connector/AudienceFeedConnectorRequestInterface"
-export * from "./interfaces/mediarithmics/api/audience-feed-connector/AudienceFeedConnectorPluginResponseInterface"
-export * from "./class/mediarithmics/audience-feed-connector/AudienceFeedConnectorBasePlugin";
-
