--- src/core/interfaces/mediarithmics/plugin/InstanceContextInterface.ts
@@ -1,9 +1,0 @@
-import {
-    AudienceFeed,
-    PluginProperty
-} from "../../../index";
-
-export interface AudienceFeedConnectorBaseInstanceContext {
-    feed: AudienceFeed;
-    feedProperties: PluginProperty[];
-}
