+++ src/mediarithmics/index.ts
@@ -2,16 +2,20 @@
 
 export * from './api/core/bidoptimizer/BidOptimizerInterface';
 export * from './api/core/common/Response';
+export * from './api/core/audiencesegment/AudienceSegmentInterface';
 export * from './api/core/creative';
 export * from './api/core/emailrouter/EmailRouterInterface';
 export * from './api/core/plugin/PluginPropertyInterface';
 export * from './api/core/plugin/ValueInterface';
 export * from './api/datamart/UserActivityInterface';
+export * from './api/plugin/audiencefeedconnector/AudienceFeedConnectorPluginResponseInterface';
+export * from './api/plugin/audiencefeedconnector/AudienceFeedConnectorRequestInterface';
 export * from './api/plugin/bidoptimizer/BidDecision';
 export * from './api/plugin/bidoptimizer/BidOptimizerRequestInterface';
 export * from './api/plugin/emailtemplaterenderer';
 export * from './api/plugin/emailtemplaterouter/EmailRouterRequestInterface';
 export * from './api/reference/UserIdentifierInterface';
+export * from './plugins/audience-feed-connector/AudienceFeedConnectorBasePlugin';
 export * from './plugins/activity-analyzer/ActivityAnalyzerBasePlugin';
 export * from './plugins/activity-analyzer/ActivityAnalyzerInterface';
 export * from './plugins/ad-renderer/base/AdRendererBasePlugin';
@@ -19,6 +23,7 @@ export * from './plugins/ad-renderer/base/AdRendererInterface';
 export * from './plugins/ad-renderer/reco-template/AdRendererRecoTemplatePlugin';
 export * from './plugins/ad-renderer/reco-template/RecommenderInterface';
 export * from './plugins/ad-renderer/template/AdRendererTemplatePlugin';
+export * from './plugins/ad-renderer/template/UserCampaignInterface';
 export * from './plugins/ad-renderer/template/UserCampaignInterface';
 export * from './plugins/bid-optimizer/BidOptimizerBasePlugin';
 export * from './plugins/common';
